NCI-Agency/anet#3150: Make calendar widget generic
Add utilities to transform calendar events from objects
